+++ src/sections/album-fav-browser.ts
@@ -141,7 +141,7 @@ export class AlbumFavBrowser extends FavBrowserBase {
 
             // load media list results.
             this.mediaList = GetAlbums(result);
-            this.mediaListLastUpdatedOn = result.lastUpdatedOn || (Date.now() / 1000);
+            this.mediaListLastUpdatedOn = result.date_last_refreshed || (Date.now() / 1000);
 
             // call base class method, indicating media list update succeeded.
             super.updatedMediaListOk();