+++ src/sections/episode-fav-browser.ts
@@ -140,7 +140,7 @@ export class EpisodeFavBrowser extends FavBrowserBase {
 
             // load media list results.
             this.mediaList = GetEpisodes(result);
-            this.mediaListLastUpdatedOn = result.lastUpdatedOn || (Date.now() / 1000);
+            this.mediaListLastUpdatedOn = result.date_last_refreshed || (Date.now() / 1000);
 
             // call base class method, indicating media list update succeeded.
             super.updatedMediaListOk();