+++ src/sections/recent-browser.ts
@@ -139,7 +139,7 @@ export class RecentBrowser extends FavBrowserBase {
 
             // load media list results.
             this.mediaList = GetTracks(result);
-            this.mediaListLastUpdatedOn = result.lastUpdatedOn || (Date.now() / 1000);
+            this.mediaListLastUpdatedOn = result.date_last_refreshed || (Date.now() / 1000);
 
             // call base class method, indicating media list update succeeded.
             super.updatedMediaListOk();