+++ src/types/spotifyplus/album-page-saved.ts
@@ -16,13 +16,6 @@ export interface IAlbumPageSaved extends IPageObject {
   items: Array<IAlbumSaved>;
 
 
-  /** 
-   * Date and time (in epoch format) of when the list was last updated.  
-   * Note that this attribute does not exist in the service response.  It was added here for convenience.
-   */
-  lastUpdatedOn?: number;
-
-
   /**
    * Gets a list of all albums contained in the underlying `Items` list.
    *