+++ src/types/spotifyplus/album-page-simplified.ts
@@ -15,12 +15,6 @@ export interface IAlbumPageSimplified extends IPageObject {
   items: Array<IAlbumSimplified>;
 
 
-  /** 
-   * Date and time (in epoch format) of when the list was last updated.  
-   * Note that this attribute does not exist in the service response.  It was added here for convenience.
-   */
-  lastUpdatedOn?: number;
-
 }