+++ src/types/spotifyplus/artist-page.ts
@@ -15,10 +15,4 @@ export interface IArtistPage extends IPageObject {
   items: Array<IArtist>;
 
 
-  /** 
-   * Date and time (in epoch format) of when the list was last updated.  
-   * Note that this attribute does not exist in the service response.  It was added here for convenience.
-   */
-  lastUpdatedOn?: number;
-
 }