+++ src/types/spotifyplus/episode-page-simplified.ts
@@ -15,13 +15,6 @@ export interface IEpisodePageSimplified extends IPageObject {
   items: Array<IEpisodeSimplified>;
 
 
-  /** 
-   * Date and time (in epoch format) of when the list was last updated.  
-   * Note that this attribute does not exist in the service response.  It was added here for convenience.
-   */
-  lastUpdatedOn?: number;
-
-
   /**
    * Checks the `Items` collection to see if an item already exists with the
    * specified Id value.