+++ src/types/spotifyplus/play-history.ts
@@ -33,10 +33,4 @@ export interface IPlayHistory {
   track: ITrack;
 
 
-  /**
-   * Date and time (in epoch format) of when the list was last updated.  
-   * Note that this attribute does not exist in the service response.  It was added here for convenience.
-   */
-  lastUpdatedOn?: number;
-
 }