+++ src/types/spotifyplus/player-queue-info.ts
@@ -36,6 +36,6 @@ export interface IPlayerQueueInfo {
    * Date and time (in epoch format) of when the list was last updated.  
    * Note that this attribute does not exist in the service response.  It was added here for convenience.
    */
-  lastUpdatedOn?: number;
+  date_last_refreshed?: number;
 
 }