+++ src/types/spotifyplus/show-page-saved.ts
@@ -17,13 +17,6 @@ export interface IShowPageSaved extends IPageObject {
   items: Array<IShowSaved>;
 
 
-  /** 
-   * Date and time (in epoch format) of when the list was last updated.  
-   * Note that this attribute does not exist in the service response.  It was added here for convenience.
-   */
-  lastUpdatedOn?: number;
-
-
   /**
    * Gets a list of all shows contained in the underlying `Items` list.
    *