+++ src/types/spotifyplus/spotify-connect-devices.ts
@@ -29,13 +29,6 @@ export interface ISpotifyConnectDevices {
   ItemsCount: number;
 
 
-  /** 
-   * Date and time (in epoch format) of when the list was last updated.  
-   * Note that this attribute does not exist in the service response.  It was added here for convenience.
-   * Used by SpotifyPlusCard only.
-   */
-  lastUpdatedOn?: number;
-
 }